# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ASPICE Documentation Platform API",
    "description": "Multi-tenant, role-based SaaS API for automating automotive ASPICE / ISO 26262 documentation from C++ source code repositories.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/projects/{project_id}/documents/stats": {
      "get": {
        "tags": ["documents"],
        "summary": "Get document status counts for a project",
        "operationId": "documentStats",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          {
            "name": "version_id",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Document stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stats": {
                      "type": "object",
                      "additionalProperties": { "type": "integer" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents": {
      "get": {
        "tags": ["documents"],
        "summary": "List documents with optional filters",
        "operationId": "listDocuments",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "name": "version_id", "in": "query", "schema": { "type": "string" } },
          { "name": "process", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "assignee_id", "in": "query", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PerPage" }
        ],
        "responses": {
          "200": {
            "description": "Paginated document list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/DocumentView" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents/approve-all": {
      "post": {
        "tags": ["documents"],
        "summary": "Approve all documents for a version (admin only)",
        "operationId": "approveAll",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ApproveAllRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk approval result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "approved_count": { "type": "integer" },
                    "document_ids": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents/assignments/batch": {
      "post": {
        "tags": ["documents"],
        "summary": "Batch assign reviewers to multiple documents (admin only)",
        "operationId": "batchAssign",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchAssignRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Batch assignment complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents/export-all": {
      "post": {
        "tags": ["documents"],
        "summary": "Trigger bulk DOCX export for a version",
        "operationId": "exportAll",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExportAllRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed download URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "download_url": { "type": "string" },
                    "expires_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents/{doc_id}": {
      "get": {
        "tags": ["documents"],
        "summary": "Get document with sections",
        "operationId": "getDocument",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/DocId" }
        ],
        "responses": {
          "200": {
            "description": "Document with sections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "document": { "$ref": "#/components/schemas/DocumentDetailView" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["documents"],
        "summary": "Update document status or due date (admin only)",
        "operationId": "updateDocument",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/DocId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateDocumentRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "document": { "$ref": "#/components/schemas/DocumentView" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents/{doc_id}/download": {
      "get": {
        "tags": ["documents"],
        "summary": "Download document as DOCX",
        "operationId": "downloadDocument",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/DocId" }
        ],
        "responses": {
          "200": {
            "description": "DOCX file",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents/{doc_id}/export": {
      "get": {
        "tags": ["documents"],
        "summary": "Export document as DOCX",
        "operationId": "exportDocument",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/DocId" }
        ],
        "responses": {
          "200": {
            "description": "DOCX file",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/documents/{doc_id}/assignments/{user_id}": {
      "delete": {
        "tags": ["documents"],
        "summary": "Remove a reviewer assignment (admin only)",
        "operationId": "removeAssignee",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/DocId" },
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Assignee removed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Project identifier"
      },
...
      "Page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "default": 1, "minimum": 1 }
      },
      "PerPage": {
        "name": "per_page",
        "in": "query",
        "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "per_page": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
...
      "ApproveAllRequest": {
        "type": "object",
        "required": ["version_id"],
        "properties": {
          "version_id": { "type": "string" },
          "process_filter": {
            "type": "array",
            "items": { "type": "string" },
            "nullable": true
          }
        }
      },
...
    }
  }
}